add WorksAnalysis to images index
