+++ common/elasticsearch/src/main/scala/uk/ac/wellcome/elasticsearch/WorksIndexConfig.scala
@@ -10,15 +10,6 @@ import com.sksamuel.elastic4s.ElasticDsl.{
   textField,
   tokenCountField
 }
-import com.sksamuel.elastic4s.json.{XContentBuilder, XContentFactory}
-import com.sksamuel.elastic4s.requests.analysis.{
-  Analysis,
-  CustomAnalyzer,
-  PathHierarchyTokenizer,
-  ShingleTokenFilter,
-  StemmerTokenFilter,
-  TokenFilter
-}
 import com.sksamuel.elastic4s.requests.mappings.{
   FieldDefinition,
   ObjectField,
@@ -26,85 +17,9 @@ import com.sksamuel.elastic4s.requests.mappings.{
 }
 import com.sksamuel.elastic4s.requests.mappings.dynamictemplate.DynamicMapping
 
-// TODO: Patch back to elastic4s
-case class AsciiFoldingTokenFilter(override val name: String,
-                                   preserveOriginal: Option[Boolean] = None)
-    extends TokenFilter {
-
-  override def build: XContentBuilder = {
-    val b = XContentFactory.jsonBuilder()
-    b.field("type", "asciifolding")
-    preserveOriginal.foreach(b.field("preserve_original", _))
-    b
-  }
-}
-
 case object WorksIndexConfig extends IndexConfig {
-  // Analysis
-  val pathTokenizer = PathHierarchyTokenizer("path_hierarchy_tokenizer")
-
-  val pathAnalyzer =
-    CustomAnalyzer("path_hierarchy_analyzer", pathTokenizer.name, Nil, Nil)
-
-  val asciiFoldingTokenFilter = AsciiFoldingTokenFilter(
-    "asciifolding_token_filter",
-    preserveOriginal = Some(true)
-  )
-
-  val shingleTokenFilter = ShingleTokenFilter(
-    "shingle_token_filter",
-    minShingleSize = Some(2),
-    maxShingleSize = Some(4))
-
-  val englishStemmerTokenFilter =
-    StemmerTokenFilter("english_token_filter", lang = "english")
-
-  val englishPossessiveStemmerTokenFilter =
-    StemmerTokenFilter(
-      "english_possessive_token_filter",
-      lang = "possessive_english")
-
-  val asciifoldingAnalyzer = CustomAnalyzer(
-    "shingle_asciifolding_analyzer",
-    tokenizer = "standard",
-    tokenFilters =
-      List("lowercase", shingleTokenFilter.name, asciiFoldingTokenFilter.name),
-    charFilters = Nil
-  )
-
-  val englishAnalyzer = CustomAnalyzer(
-    "english_analyzer",
-    tokenizer = "standard",
-    tokenFilters = List(
-      "lowercase",
-      asciiFoldingTokenFilter.name,
-      englishStemmerTokenFilter.name,
-      englishPossessiveStemmerTokenFilter.name),
-    charFilters = Nil
-  )
-
-  val shingleAsciifoldingAnalyzer = CustomAnalyzer(
-    "shingle_asciifolding_analyzer",
-    tokenizer = "standard",
-    tokenFilters =
-      List("lowercase", shingleTokenFilter.name, asciiFoldingTokenFilter.name),
-    charFilters = Nil
-  )
-
-  val analysis = Analysis(
-    analyzers = List(
-      pathAnalyzer,
-      asciifoldingAnalyzer,
-      shingleAsciifoldingAnalyzer,
-      englishAnalyzer
-    ),
-    tokenFilters = List(
-      asciiFoldingTokenFilter,
-      shingleTokenFilter,
-      englishStemmerTokenFilter,
-      englishPossessiveStemmerTokenFilter),
-    tokenizers = List(pathTokenizer)
-  )
+  import WorksAnalysis._
+  val analysis = WorksAnalysis()
 
   // Fields
   val sourceIdentifier = objectField("sourceIdentifier")
